Handle UserKit script in a separate controller
